tests/unit/requirements.txt
PyYAML
pytest
pytest-cov
pytest-subtests
